refactor(input): share fuzzy search scoring
