+++ src/input/state/core/command_palette/mod.rs
@@ -18,7 +18,7 @@ pub(crate) use layout::{
 pub use registry::{CommandEntry, command_palette_entries};
 pub use search::CommandPaletteListRow;
 pub(in crate::input::state::core) use search::CommandPaletteResults;
-pub(crate) use search::{action_meta_token_score, fuzzy_score, query_tokens};
+pub(crate) use search::{action_meta_token_score, query_tokens};
 
 /// Cursor hint for different regions of the command palette.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
